+++ packages/reaxes-vue2/src/reaxper.tsx
@@ -4,3 +4,4 @@ export const reaxper = <T extends {}>(component:T):T => {
 
 
 import { observer } from 'reaxes-vue2/libs/mobx-vue';
+import type {Component} from 'vue2/types';