+++ packages/reaxes/public/orzMobx.d.ts
@@ -1,5 +1,6 @@
-export declare const orzMobx: <S extends object>(state: S) => {
+export declare const orzMobx: <S extends {}>(state: S) => {
     store: S;
-    setState: <T extends Partial<S>>(partialState : T) => S;
+    setState: <state extends Partial<S>>(partialState: state) => S;
+	 mutate : <T extends (store:S) => void>(fn:T) => void;
     mutatePartialState: <T extends object = S>(partialState: Partial<T>, deepStore?: any) => void;
 };